fix(publishing): changed group name to io.github.flaxoos to match new repo name. now the group name is owned by me and so plugins can be published to gradle portal.
Also added shadow jar to include dependencies
