--- plugins/common/src/main/kotlin/io/flax/KoverIgnore.kt
@@ -1,5 +1,0 @@
-package io.flax
-
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class KoverIgnore
