+++ plugins/common/src/main/kotlin/io/github/flaxoos/KoverIgnore.kt
@@ -1,0 +1,5 @@
+package io.github.flaxoos
+
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class KoverIgnore
