--- plugins/kover-badge-plugin/gradle.properties
@@ -1,9 +1,0 @@
-ID=io.github.flax.kover-badge
-VERSION=0.0.4
-GROUP=io.github.flax
-DISPLAY_NAME=Kover Badge Plugin
-DESCRIPTION=A plugin to generate and coverage badges for Kotlin Kover reports and embed them in readme files
-WEBSITE=https://github.com/idoflax/kover-badge-plugin"
-VCS_URL=https://github.com/idoflax/kover-badge-plugin"
-IMPLEMENTATION_CLASS=io.flax.kover.KoverBadgePlugin
-github.repository.name=flax-gradle-plugins
