+++ plugins/kover-badge/gradle.properties
@@ -1,0 +1,9 @@
+ID=io.github.flaxoos.kover-badge
+VERSION=0.0.5
+GROUP=io.github.flaxoos
+DISPLAY_NAME=Kover Badge Plugin
+DESCRIPTION=A plugin to generate and coverage badges for Kotlin Kover reports and embed them in readme files
+WEBSITE=https://github.com/flaxoos/kover-badge-plugin"
+VCS_URL=https://github.com/flaxoos/kover-badge-plugin"
+IMPLEMENTATION_CLASS=io.github.flaxoos.kover.KoverBadgePlugin
+github.repository.name=flax-gradle-plugins
