+++ plugins/kover-badge/src/functionalTest/resources/test-project-files/gitignore
