+++ test-projects/kover-badge/src/main/kotlin/io/github/flaxoos/kover/Application.kt
@@ -1,0 +1,20 @@
+package io.github.flaxoos.kover
+
+class Application(private val helloSayer: HelloSayer, private val shouldSayGoodbye: Boolean) {
+    fun run() {
+        helloSayer.sayHello()
+        if (shouldSayGoodbye) {
+            helloSayer.sayGoodbye()
+        }
+    }
+}
+
+class HelloSayer(private val name: String) {
+    fun sayHello() {
+        println("Hello $name")
+    }
+
+    fun sayGoodbye() {
+        println("Goodbye $name")
+    }
+}
